Update SRL/core/flag.simba
Strangely, the RS minimap is not a perfect circle, as radius from centre to bottom is only 75 while that of to other edges is 76.
The minimum radius should be set at 75. I've thought it was reduced to ensure flag will be on screen but for that to happen when clicking top parts of MM, the radius has to be reduced to ~60.
So if anyone clicking on the edge of MM wishes to use flag functions they should set a time allowance for FlagPresent. (as it will appear when you walk closer)
The integrity of the function should be maintained in that points beyond 73 radius and up to 75 should not return false.