Fixes RadialWalk and friends.
Fixes RadialWalkEx to enable search in the counterclockwise direction,
like is described in picture 1 in this post:
http://villavu.com/forum/showthread.php?t=372&p=3370#post3370

Does this by bypassing the FixD inside of variantToDirection, as
negative degrees are already fixed in FilterPointsPie.

This commit also fixes the ability to pass compass directions to RadialWalkEx,
ex. 'n', 's', etc.

Signed-off-by: idynin <ilya@idynin.com>